@@ -1,9 +1,9 @@
 source 'https://rubygems.org'
 
-gem 'dotenv-deployment'
+gem 'dotenv-rails'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.7.1'
+gem 'rails', '5.2.0'
 
 gem 'pg'
 
@@ -19,7 +19,7 @@ group :assets do
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   gem 'therubyracer', platforms: :ruby
 
-  gem 'uglifier', '>= 1.3.0'
+  gem 'uglifier'
   gem "less-rails"
 end
 
@@ -34,7 +34,7 @@ gem 'sinatra', :require => nil
 gem 'foreman'
 gem 'haml-rails'
 # Use pull request that has needed Rails 4 improvements https://github.com/pkurek/flatui-rails/pull/25
-gem 'flatui-rails', github: 'iffyuva/flatui-rails', ref: '3d3c423'
+gem 'flatui-rails', git: 'https://github.com/iffyuva/flatui-rails.git', ref: '3d3c423'
 gem 'font-awesome-rails'
 gem "file-tail"
 gem 'syslog_protocol'
@@ -46,20 +46,19 @@ gem 'gravatar_image_tag'
 gem "formtastic"
 # Need commit c9331088146e456a69bd6e94298c80d09be3ee74
 gem 'formtastic-bootstrap', git: "https://github.com/mjbellantoni/formtastic-bootstrap.git", ref: "f86eaef93bea0a06879b3977d7554864964a623f"
-gem 'factory_girl_rails'
+gem 'factory_bot_rails'
 gem 'haml-coderay'
 gem 'nokogiri'
 gem 'google-analytics-rails'
 gem 'premailer'
-gem "archive-tar-minitar"
+gem "minitar"
 gem "pundit"
 gem "friendly_id"
 gem "user_agent_parser"
-gem "mail_form"
 gem 'newrelic_rpm'
 gem 'honeybadger'
 gem 'dkim'
-gem 'fog'
+gem 'fog-aws'
 
 # To use ActiveModel has_secure_password
 # gem 'bcrypt-ruby', '~> 3.0.0'
@@ -88,8 +87,7 @@ group :development do
   gem "spring-commands-rspec"
   gem "capistrano", "~> 2"
   gem 'rvm-capistrano', ">= 1.5.6", require: false
-  # Newer versions don't support Ruby 2.1. Remove this when we upgrade Ruby
-  gem "listen", "~> 2"
+  gem "listen"
 end
 
 group :test do
@@ -98,6 +96,7 @@ group :test do
   gem 'climate_control'
   gem "vcr"
   gem "webmock"
+  gem 'rails-controller-testing'
 end
 
 group :development, :test do
