@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :app do
     team
     name "My App"
