@@ -70,11 +70,11 @@ describe "getting a bunch of screenshots", js: true do
         @email = @app.emails.create!(from: "hello@cuttlefish.io", to: "matthew@openaustralia.org", data: mail.encoded)
         @delivery = @email.deliveries.first
         @delivery.update_attributes(sent: true, open_tracked: true)
-        FactoryGirl.create(:postfix_log_line, delivery: @delivery,
+        FactoryBot.create(:postfix_log_line, delivery: @delivery,
             time: 5.minutes.ago, dsn: "2.0.0", extended_status: "sent (250 2.0.0 r6ay1l02Y4aTF9m016ayyY mail accepted for delivery)")
-        FactoryGirl.create(:open_event, delivery: @delivery, created_at: 2.minutes.ago, user_agent: "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_7_5) AppleWebKit/537.31 (KHTML, like Gecko) Chrome/26.0.1410.65 Safari/537.31",
+        FactoryBot.create(:open_event, delivery: @delivery, created_at: 2.minutes.ago, user_agent: "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_7_5) AppleWebKit/537.31 (KHTML, like Gecko) Chrome/26.0.1410.65 Safari/537.31",
           ip: "1.2.3.4")
-        FactoryGirl.create(:delivery_link, delivery_id: @delivery.id)
+        FactoryBot.create(:delivery_link, delivery_id: @delivery.id)
 
       visit new_admin_session_path
       fill_in "Email", with: "matthew@openaustralia.org"
