@@ -4,7 +4,7 @@ describe OutgoingDelivery do
   describe "#send" do
     context "an email with one recipient" do
       before :each do
-        @email = FactoryGirl.create(:email, to: "foo@bar.com", data: "from: contact@foo.com\nto: foo@bar.com\n\nMy original data")
+        @email = FactoryBot.create(:email, to: "foo@bar.com", data: "from: contact@foo.com\nto: foo@bar.com\n\nMy original data")
         @outgoing = OutgoingDelivery.new(@email.deliveries.first)
       end
 
