@@ -5,9 +5,9 @@ describe Address do
     let (:address1) { Address.create!(text: "matthew@foo.com") }
     let (:address2) { Address.create!(text: "peter@bar.com") }
     before :each do
-      @email1 = FactoryGirl.create(:email, from_address: address1, to_addresses: [address1])
-      @email2 = FactoryGirl.create(:email, from_address: address1, to_addresses: [address2])
-      @email3 = FactoryGirl.create(:email, from_address: address2, to_addresses: [address2])
+      @email1 = FactoryBot.create(:email, from_address: address1, to_addresses: [address1])
+      @email2 = FactoryBot.create(:email, from_address: address1, to_addresses: [address2])
+      @email3 = FactoryBot.create(:email, from_address: address2, to_addresses: [address2])
     end
 
     describe "#emails_sent" do
