@@ -5,16 +5,14 @@ class DeliveriesController < ApplicationController
   def index
     @status = params[:status]
     @search = params[:search]
+    @deliveries = policy_scope(Delivery)
     if params[:app_id]
       @app = App.find(params[:app_id])
-      authorize @app, :show?
-      @deliveries = @app.deliveries
-    else
-      @deliveries = policy_scope(Delivery)
+      @deliveries = @deliveries.where(app_id: @app.id)
+      @deliveries = @deliveries.joins(:email).where("emails.app_id" => @app.id)
     end
 
     @deliveries = @deliveries.where(status: @status) if @status
-    @deliveries = @deliveries.joins(:email).where("emails.app_id" => @app.id) if @app
     @deliveries = @deliveries.joins(:address).where("addresses.text" => @search) if @search
 
     @deliveries = @deliveries.includes(:delivery_links, :postfix_log_lines, :email, :address).order("deliveries.created_at DESC").page(params[:page])
