@@ -21,7 +21,7 @@ describe Filters::InlineCss do
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN" "http://www.w3.org/TR/REC-html40/loose.dtd">
 <html>
 <head><meta http-equiv="Content-Type" content="text/html; charset=UTF-8"></head>
-<body><p style="font-size: 20px">This is HTML with “some” UTF-8</p></body>
+<body><p style="font-size: 20px;">This is HTML with “some” UTF-8</p></body>
 </html>
       EOF
     end
